requirements.txt
flask==0.10.1
flask-restful==0.2.12
Flask-HTTPAuth==1.0.0
tornado==4.0.1
subprocess32==3.2.6
Flask-SocketIO
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
gevent==1.0.1
gevent-socketio
gevent-websocket==0.9.2
greenlet==0.4.2
itsdangerous==0.23
ujson==1.33
ivy-python
